Add files via upload
Moka Store is a modern e-commerce website designed to sell clothing, shoes, and accessories for both men and women. It features a clean, responsive design and user-friendly layout, making shopping easy and enjoyable. The front-end was built with HTML, CSS, and JavaScript, focusing on speed, accessibility, and a smooth user experience.